+++ sdk/cosmos/azure-cosmos/azure/cosmos/_execution_context/base_execution_context.py
@@ -132,7 +132,6 @@ class _QueryExecutionContextBase(object):
             else:
                 self._continuation = None
             if fetched_items:
-                self._client.diagnostics.update_diagnostics(response_headers, fetched_items)
                 break
         return fetched_items