+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_user.py
@@ -317,18 +317,3 @@ class UserProxy(object):
         if response_hook:
             response_hook(self.client_connection.last_response_headers, result)
 
-    async def diagnostics(self, p=False):
-        """Returns a dictionary of the diagnostics from the last request.
-
-        Best used when catching an exception.
-
-        Ex:
-            >>>try:
-            >>>    database = client.create_database_if_not_exists(id="DatabaseTest")
-            >>>except:
-            >>>    client.diagnostics(p=True)
-
-        :param p: Defaulted to False, if set to true will print the diagnostics in a nicely formatted style.
-        :rtype: dict
-        """
-        return self.client_connection.diagnostics(p=p)